+++ cli/commands/silenced/info_test.go
@@ -86,40 +86,31 @@ func TestInfoCommandRunEClosureWithErr(t *testing.T) {
 	assert.Empty(out)
 }
 
-func Test_expireTime(t *testing.T) {
+func Test_expireAt(t *testing.T) {
 	unixToInternal := int64((1969*365 + 1969/4 - 1969/100 + 1969/400) * 24 * 60 * 60)
 	future := time.Unix(1<<63-1-unixToInternal, 999999999)
 
 	tests := []struct {
-		name          string
-		beginTS       int64
-		expireSeconds int64
-		want          string
+		name     string
+		expireAt int64
+		want     string
 	}{
 		{
-			name:          "an entry without an expiration returns -1",
-			beginTS:       time.Now().Truncate(time.Duration(1) * time.Minute).Unix(),
-			expireSeconds: -1,
-			want:          "-1",
+			name:     "an entry without an expiration returns -1",
+			expireAt: -1,
+			want:     "no expiration",
 		},
 		{
-			name:          "an entry that is not yet in effect return a RFC3339 date",
-			beginTS:       future.Unix(),
-			expireSeconds: 300,
-			want:          "292277024627-12",
-		},
-		{
-			name:          "an entry that is in effect return the configured duration",
-			beginTS:       time.Now().Truncate(time.Duration(1) * time.Minute).Unix(),
-			expireSeconds: 300,
-			want:          "5m",
+			name:     "an RFC3339 date is returned",
+			expireAt: future.Unix(),
+			want:     "292277024627-12",
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got := expireTime(tt.beginTS, tt.expireSeconds)
+			got := expireAt(tt.expireAt)
 			if !strings.Contains(got, tt.want) {
-				t.Errorf("expireTime() = %v, want %v", got, tt.want)
+				t.Errorf("expireAt() = %v, want %v", got, tt.want)
 			}
 		})
 	}